+++ backend/package.json
@@ -11,6 +11,7 @@
     "express": "^5.2.1",
     "express-oauth2-jwt-bearer": "^1.7.4",
     "jsonwebtoken": "^9.0.3",
+    "jwks-rsa": "^4.0.1",
     "mongodb": "^7.1.0",
     "mongoose": "^9.3.0",
     "socket.io": "^4.8.3"
